+++ changelog.md
@@ -1,4 +1,6 @@
 # Changelog
 
-**./test-parts/readme.md**
-* Deleted placeholder file.
+**./test-parts/test-graph-parse.js**
+* Renamed the following:
+	* 'handleValidCases' function to 'handleValidCase'
+	* "Parse Graph" - "Valid Cases" section to "Valid Case"